requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyaml==16.9.0
PyYAML==3.12
requests==2.11.1
six==1.10.0
slackclient==1.0.2
websocket-client==0.37.0
Werkzeug==0.11.11
python-dotenv==0.7.1
pyppeteer==0.0.12
tinys3==0.1.12
